requirements.txt

openai
chromadb
tiktoken
faiss-cpu
streamlit
pyyaml
watchdog
sentence-transformers
langchain-community
